refactor: 회원가입 관련 파일 구조 개선
