+++ src/app/api/API.tsx
@@ -1,30 +1,9 @@
 import axios, { AxiosResponse } from "axios";
 import { GetScheduleQuery, Schedule } from "@type/schedule.tsx";
-import { ServerState, SignInterface, SignUp, User } from "@type/common.tsx";
+import { ServerState, SignInterface, User } from "@type/common.tsx";
 import { url } from "./url.ts";
 import { getSessionStorage, setSessionStorage } from "../utils/storage.ts";
 import { LOCAL_STORAGE_KEY_SERVER } from "./keys.ts";
-
-/**
- * 반드시 서버로 요청할 때 객체 형식이어야 JSON 으로 변환되어 서버에 잘 들어감!
- * 아직 타입이 미정인 코드들이 있어서 일단 any 처리
- */
-
-export const fetchSignUp = async (user: SignUp) => {
-  try {
-    const server = getSessionStorage<ServerState>(
-      LOCAL_STORAGE_KEY_SERVER,
-      "real"
-    );
-    const response = await axios.post<boolean>(
-      `${url[server]}/fin-the-pen-web/sign-up`,
-      user
-    );
-    return response.data;
-  } catch (err) {
-    alert(err);
-  }
-};
 
 export const fetchLogin = async (sign: SignInterface) => {
   try {
