+++ src/pages/SignUp/Header.tsx
@@ -1,5 +1,5 @@
-import { Stack, Typography } from '@mui/material';
-import LogoCircle from '../../../components/common/LogoCircle';
+import { Stack, Typography } from "@mui/material";
+import LogoCircle from "@components/common/LogoCircle.tsx";
 
 function Header() {
   return (
