+++ src/pages/SignUp/SignUp.tsx
@@ -1,19 +1,15 @@
-import { Stack } from '@mui/material';
-import CenterBox from '../../components/layouts/CenterBox';
-import useHeader from '../../hooks/useHeader';
-import Header from '../../containers/sign/SignUpContainer/Header';
-import SignUpFields from '../../containers/sign/SignUpContainer/SignUpFields';
+import { Stack } from "@mui/material";
+import CenterBox from "../../components/layouts/CenterBox";
+import useHeader from "../../hooks/useHeader";
+import Header from "./Header.tsx";
+import SignUpFields from "./SignUpFields.tsx";
 
 function SignUp() {
   useHeader(false);
 
   return (
     <CenterBox>
-      <Stack
-        justifyContent="center"
-        alignItems="center"
-        px={1}
-      >
+      <Stack justifyContent="center" alignItems="center" px={1}>
         <Header />
         <SignUpFields />
       </Stack>
